pipeline_docker-compose/slack_bot/requirements.txt
pyjokes
slackclient
SQLAlchemy
psycopg2-binary